Add commands for biuld flankExample and earlGreyExample
